CALARTS / Coursera - Class # 2 - Work for week 1 and 2
This is for the CALARTS / Coursera - Class # 2 for UI/UX- Work for week 1 and Week 2. Week one was coming up with an application idea. Week two was getting more into the UX of it and creating a Site Map.

It was nice to wireframe and get some real UX going.